requirements.txt
fastapi
supabase
uvicorn
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
opencv-python
numpy
openai
langchain>=0.1.0
langgraph
tiktoken
python-dotenv
langchain-community
langchain-groq
motor
pymongo
bcrypt
langchain
langgraph
langchain-community
langchain-core
python-pptx
